Move `Overview.md` to `README.md`
This change makes renders the overview for folks visiting this repo on github.

While `Overview.md` is typically used for overviews of Wasm proposals, it
doesn't work well for this benchmarks repo:

* The proposals already have a `README.md`, which typically links to the
rendered proposal on github pages and to the overview document. Github visitors
to a proposal's repo aren't faced with a blank listing of files/directories.

* This repo isn't a proposal for Wasm, like (say) reference types, and is more
of a resource for the Wasm ecosystem, like (say) WABT. WABT doesn't have an
`Overview.md`, it has a `README.md`.